+++ packages/core/src/index.ts
@@ -40,6 +40,7 @@ export * from "./hooks/UseMultiOnClickOutside";
 export * from "./hooks/UseOnClickOutside";
 export * from "./hooks/UseOnNoMouseInput";
 export * from "./hooks/UseForwardedRef";
+export * from "./hooks/UseTimeoutState";
 
 export * from "./utils/SwitchCaseExhauster";
 export * from "./utils/parsers/NumberParser";