+++ packages/elements/src/components/ui/action-menu/ActionMenuItem.tsx
@@ -1,141 +1,79 @@
-import { IconDefinition } from "@fortawesome/fontawesome-svg-core";
-import { Row, Space, Text, Box } from "@stenajs-webui/core";
+import { ButtonElementProps, Row } from "@stenajs-webui/core";
 import * as React from "react";
-import { forwardRef, PropsWithChildren, useCallback, useContext } from "react";
-import { Icon } from "../icon/Icon";
-import { ActionMenuContext } from "./ActionMenuContext";
+import { forwardRef } from "react";
 import cx from "classnames";
-import { InputSpinner } from "../spinner/InputSpinner";
-import { useFocusManager } from "@react-aria/focus";
 
 import styles from "./ActionMenuItem.module.css";
+import {
+  ButtonContent,
+  ButtonContentProps,
+} from "../buttons/common/ButtonContent";
+import { useActionMenuLogic } from "./UseActionMenuLogic";
 
-export type ActionMenuItemVariant = "standard" | "danger";
+export type ActionMenuItemVariant = "standard" | "danger" | "success";
 
-export interface ActionMenuItemProps {
-  id?: string;
-  label: string;
+export interface ActionMenuItemProps
+  extends ButtonElementProps,
+    ButtonContentProps {
   variant?: ActionMenuItemVariant;
-  rightText?: string;
-  icon?: IconDefinition;
-  iconRight?: IconDefinition;
-  disabled?: boolean;
   disableCloseOnClick?: boolean;
-  loading?: boolean;
-  onClick?: () => void;
 }
 
 export const ActionMenuItem = forwardRef<
   HTMLButtonElement,
-  PropsWithChildren<ActionMenuItemProps>
+  ActionMenuItemProps
 >(function ActionMenuItem(
   {
-    label,
-    variant = "standard",
-    icon,
-    iconRight,
-    rightText,
-    disabled,
-    onClick,
-    children,
-    disableCloseOnClick,
+    success,
     loading,
-    id,
+    leftIcon,
+    left,
+    right,
+    rightIcon,
+    label,
+    iconClassName,
+    labelClassName,
+    spinnerClassName,
+    leftWrapperClassName,
+    variant = "standard",
+    className,
+    ...props
   },
   ref
 ) {
-  const { close } = useContext(ActionMenuContext);
-
-  const onClickHandler = useCallback(() => {
-    if (close && !disableCloseOnClick) {
-      close();
-    }
-    if (onClick) {
-      onClick();
-    }
-  }, [onClick, close, disableCloseOnClick]);
-
-  const focusManager = useFocusManager();
-  const onKeyDown = (event: React.KeyboardEvent<HTMLButtonElement>) => {
-    switch (event.key) {
-      case "ArrowDown":
-      case "ArrowRight":
-        event.preventDefault();
-        focusManager.focusNext({ wrap: true });
-        break;
-      case "ArrowUp":
-      case "ArrowLeft":
-        event.preventDefault();
-        focusManager.focusPrevious({ wrap: true });
-        break;
-    }
-  };
+  const { onClickHandler, onKeyDown, innerRef } = useActionMenuLogic(
+    props,
+    ref
+  );
 
   return (
     <button
-      id={id}
-      className={cx(styles.actionMenuItem, styles[variant])}
+      {...props}
+      className={cx(styles.actionMenuItem, styles[variant], className)}
       onKeyDown={onKeyDown}
-      onClick={disabled ? undefined : onClickHandler}
-      disabled={disabled}
-      ref={ref}
+      onClick={props.disabled ? undefined : onClickHandler}
+      ref={innerRef}
     >
       <Row
+        alignItems={"center"}
         width={"100%"}
         indent={2}
-        alignItems={"center"}
-        justifyContent={"space-between"}
+        className={styles.actionMenuItemInnerContent}
       >
-        <Row alignItems={"center"}>
-          {loading ? (
-            <>
-              <Box width={20} alignItems={"center"}>
-                <InputSpinner />
-              </Box>
-              <Space />
-            </>
-          ) : (
-            icon && (
-              <>
-                <Icon
-                  className={styles.actionMenuItemIcon}
-                  icon={icon}
-                  fixedWidth
-                  size={16}
-                />
-                <Space />
-              </>
-            )
-          )}
-          <Text className={styles.actionMenuItemLabel} whiteSpace={"nowrap"}>
-            {label}
-          </Text>
-        </Row>
-        {rightText && (
-          <Text
-            className={styles.actionMenuItemText}
-            size={"small"}
-            whiteSpace={"nowrap"}
-          >
-            {rightText}
-          </Text>
-        )}
-        {children && (
-          <>
-            <Space />
-            {children}
-          </>
-        )}
-        {iconRight && (
-          <>
-            <Space />
-            <Icon
-              className={styles.actionMenuItemIcon}
-              icon={iconRight}
-              size={14}
-            />
-          </>
-        )}
+        <ButtonContent
+          success={success}
+          loading={loading}
+          leftIcon={leftIcon}
+          left={left}
+          right={right}
+          rightIcon={rightIcon}
+          label={label}
+          labelClassName={styles.actionMenuItemLabel}
+          iconClassName={styles.actionMenuItemIcon}
+          leftWrapperClassName={cx({
+            [styles.actionMenuItemIconWrapper]: success || loading || leftIcon,
+          })}
+        />
       </Row>
     </button>
   );